deribit: convert order book to TS
